+++ cmd/cmd_check_test.go
@@ -2,6 +2,7 @@ package cmd_test
 
 import (
 	. "github.com/sky-uk/umc-shared/vergo/internal-test"
+	"github.com/sky-uk/umc-shared/vergo/release"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -17,7 +18,7 @@ func TestCheckReleaseSuccess(t *testing.T) {
 
 func TestCheckReleaseFailure(t *testing.T) {
 	_, tempDir := PersistentRepository(t)
-	cmd, buffer := makeCheckFail(t)
+	cmd, buffer := makeCheckFail(t, release.ErrSkipRelease, success)
 	cmd.SetArgs([]string{"check", "release", "--repository-location", tempDir, "-t", "some-prefix", "--log-level", "error", "-p"})
 	err := cmd.Execute()
 	assert.NotNil(t, err)
@@ -26,20 +27,29 @@ func TestCheckReleaseFailure(t *testing.T) {
 
 func TestCheckRelease2Failures(t *testing.T) {
 	_, tempDir := PersistentRepository(t)
-	cmd, buffer := makeCheckWithCheckList(t, checkReleaseFail(t), checkReleaseFailForAnotherReason(t))
+	cmd, buffer := makeCheckFail(t, release.ErrSkipRelease, release.ErrInvalidHeadless)
 	cmd.SetArgs([]string{"check", "release", "--repository-location", tempDir, "-t", "some-prefix", "--log-level", "error", "-p"})
 	err := cmd.Execute()
 	assert.NotNil(t, err)
 	assert.Equal(t, `Error: skip release hint present
-it does not look right
+invalid headless checkout
 `, readBuffer(t, buffer))
 }
 
-func TestCheckReleaseSuccessfulAndFailingCheck(t *testing.T) {
+func TestCheckReleaseInvalidHeadless(t *testing.T) {
 	_, tempDir := PersistentRepository(t)
-	cmd, buffer := makeCheckWithCheckList(t, checkReleaseFail(t), checkReleaseSuccess(t))
+	cmd, buffer := makeCheckFail(t, success, release.ErrInvalidHeadless)
 	cmd.SetArgs([]string{"check", "release", "--repository-location", tempDir, "-t", "some-prefix", "--log-level", "error", "-p"})
 	err := cmd.Execute()
 	assert.NotNil(t, err)
-	assert.Equal(t, "Error: skip release hint present\n", readBuffer(t, buffer))
+	assert.Equal(t, "Error: invalid headless checkout\n", readBuffer(t, buffer))
+}
+
+func TestCheckReleaseUnexpectedBranch(t *testing.T) {
+	_, tempDir := PersistentRepository(t)
+	cmd, buffer := makeCheckFail(t, success, release.ErrUnexpectedBranch)
+	cmd.SetArgs([]string{"check", "release", "--repository-location", tempDir, "-t", "some-prefix", "--log-level", "error", "-p"})
+	err := cmd.Execute()
+	assert.NotNil(t, err)
+	assert.Equal(t, "Error: unexpected branch\n", readBuffer(t, buffer))
 }