+++ cmd/common_test.go
@@ -3,18 +3,18 @@ package cmd_test
 import (
 	"bytes"
 	"errors"
-	"fmt"
 	"github.com/Masterminds/semver/v3"
 	"github.com/go-git/go-git/v5"
 	. "github.com/sky-uk/umc-shared/vergo/cmd"
 	vergo "github.com/sky-uk/umc-shared/vergo/git"
 	. "github.com/sky-uk/umc-shared/vergo/internal-test"
-	"github.com/sky-uk/umc-shared/vergo/release"
 	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"testing"
 )
+
+var success error = nil
 
 func bumpSuccess(t *testing.T) BumpFunc {
 	t.Helper()
@@ -23,33 +23,21 @@ func bumpSuccess(t *testing.T) BumpFunc {
 	}
 }
 
-func mockPushTagSuccess(repo *git.Repository, socket, version, prefix, remote string, dryRun bool) error {
+func mockPushTagSuccess(_ *git.Repository, _, _, _, _ string, _ bool) error {
 	return nil
 }
 
-func mockPushTagFailure(repo *git.Repository, socket, version, prefix, remote string, dryRun bool) error {
+func mockPushTagFailure(_ *git.Repository, _, _, _, _ string, _ bool) error {
 	return errors.New("push tag failed")
 }
 
-func checkReleaseSuccess(t *testing.T) CheckReleaseFunc {
+func checkReleaseDependencies(t *testing.T, err1 error, err2 error) (SkipHintPresentFunc, ValidateHEADFunc) {
 	t.Helper()
 	return func(repo *git.Repository, tagPrefixRaw string) error {
-		return nil
-	}
-}
-
-func checkReleaseFail(t *testing.T) CheckReleaseFunc {
-	t.Helper()
-	return func(repo *git.Repository, tagPrefixRaw string) error {
-		return release.ErrSkipRelease
-	}
-}
-
-func checkReleaseFailForAnotherReason(t *testing.T) CheckReleaseFunc {
-	t.Helper()
-	return func(repo *git.Repository, tagPrefixRaw string) error {
-		return fmt.Errorf("it does not look right")
-	}
+			return err1
+		}, func(repo *git.Repository, versionedBranches []string) error {
+			return err2
+		}
 }
 
 func makeBump(t *testing.T) (*cobra.Command, *bytes.Buffer) {
@@ -75,28 +63,17 @@ func pushTagFail(t *testing.T) (*cobra.Command, *bytes.Buffer) {
 func makeCheck(t *testing.T) (*cobra.Command, *bytes.Buffer) {
 	t.Helper()
 	cmd := RootCmd()
-	cmd.AddCommand(CheckCmd([]CheckReleaseFunc{checkReleaseSuccess(t)}))
+	cmd.AddCommand(CheckCmd(checkReleaseDependencies(t, success, success)))
 	b := bytes.NewBufferString("")
 	cmd.SetOut(b)
 	cmd.SetErr(b)
 	return cmd, b
 }
 
-func makeCheckFail(t *testing.T) (*cobra.Command, *bytes.Buffer) {
+func makeCheckFail(t *testing.T, err1 error, err2 error) (*cobra.Command, *bytes.Buffer) {
 	t.Helper()
 	cmd := RootCmd()
-	cmd.AddCommand(CheckCmd([]CheckReleaseFunc{checkReleaseFail(t)}))
-	b := bytes.NewBufferString("")
-	cmd.SetOut(b)
-	cmd.SetErr(b)
-	return cmd, b
-}
-
-func makeCheckWithCheckList(t *testing.T, checks ...CheckReleaseFunc) (*cobra.Command, *bytes.Buffer) {
-	t.Helper()
-	assert.False(t, len(checks) == 0)
-	cmd := RootCmd()
-	cmd.AddCommand(CheckCmd(checks))
+	cmd.AddCommand(CheckCmd(checkReleaseDependencies(t, err1, err2)))
 	b := bytes.NewBufferString("")
 	cmd.SetOut(b)
 	cmd.SetErr(b)